# Conflicts:
#	Code/ShipWorks.Data.Model/EntityClasses/ExcludedPackageTypeEntity.cs
#	Code/ShipWorks/Data/Administration/Scripts/Update/4.3/4.3.0.0.sql
#	Code/ShipWorks/ShipWorks.Core.csproj

+++ Code/ShipWorks/Stores/Platforms/Shopify/ShopifyDownloader.cs
@@ -20,6 +20,7 @@ using ShipWorks.Users;
 using log4net;
 using ShipWorks.Common.Threading;
 using System.Diagnostics;
+using Interapptive.Shared.Business.Geography;
 
 namespace ShipWorks.Stores.Platforms.Shopify
 {
@@ -262,8 +263,8 @@ namespace ShipWorks.Stores.Platforms.Shopify
                 order.OrderDate = jsonOrder.GetValue<DateTime>("created_at", DateTime.MinValue).ToUniversalTime();
 
                 //Get the customer
-                int onlineCustomerID = jsonOrder.GetValue<int>("customer.id", -1);
-                order.OnlineCustomerID = (onlineCustomerID == -1) ? (int?)null : onlineCustomerID;
+                long onlineCustomerID = jsonOrder.GetValue<long>("customer.id", -1);
+                order.OnlineCustomerID = (onlineCustomerID == -1) ? (long?)null : onlineCustomerID;
 
                 //Requested shipping
                 if (requestedShippingField == ShopifyRequestedShippingField.Code)
